+++ src/component/UI/error/Error.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import classes from './Error.module.css';
+
 interface IErrorMessageProps {
   message: string;
 }
